# Faster Bloch GPU