@cncastillo 2025-01-24
Wow amazing! Thanks @rkierulf ! Currently the CI is a little messed up, so I cannot merge it straight away. I benchmarked the previous implementation, with this code:

[GPU_Benchmarks_ISMRM2025.zip](https://github.com/user-attachments/files/18536991/GPU_Benchmarks_ISMRM2025.zip)

And got:
![image](https://github.com/user-attachments/assets/1fd306b6-7522-491c-ab82-daeb78b6640b)

The previous method didn't scale very well. Do you think the changes you have made can improve this? There are clearly two regimes, flat and linear growth, not sure if they correspond to the memory- and compute-bound regimes. What are your thoughts on this?