@rkierulf 2025-01-24
Interesting! My guess is that the transition from flat to linear growth begins when the simulation becomes large enough to fully saturate the resources of the GPU. After that, due to the memory bound nature of the problem the growth in time should be proportional to the number of additional reads + writes to GPU global memory per added spin index. Since the new implementation has far fewer GPU memory operations per spin index, I would expect the slope to be less steep. And hopefully, it is more efficient overall so the curve would shift somewhat to the right as well. I'd be curious to see the results of this script run again with the new changes.

Regarding the CI Failures:

- OneAPI and non-Buildkite CI failures: appear to be preexisting?
- Metal failures: these are the same two tests I was referencing that were only failing in my test environment, except here they are failing on Julia 1.10. Very strange, I'll try to investigate some more
- CUDA and AMD failures: both are failing on 1.10 while calling `KA.synchronize()` in `run_sim_time_iter!`. Maybe I need to be more conservative and call `KA.synchronize()` immediately after the kernels finish, I'll see if this fixes the issue