@cncastillo 2025-01-25
Thanks for investigating. I will for rerun the benchmarks using this PR  when I have more time (moving to Stanford tomorrow, and then I have a conference in Washington DC).

I believe the problem with Julia 1.11 and oneAPI is: https://github.com/JuliaGPU/oneAPI.jl/issues/473. 

Recently they changed the internals of all the GPU packages to be based on KA.jl. onAPI.jl v2.0 should include these changes, so bumping the compat to include `2` might fix than error, but now sure. 

Don't mind the GitHub CI errors, @Stockless is currently fixing those.