@rkierulf 2025-01-27
The tests are all passing now, except for the oneAPI tests with Julia 1.11. For the 1.11 tests overall, I'm not sure it is using the local versions of the packages. With the 1.10 tests I can see in the Manifest info printed before the tests are run:

```
[d0bc0b20] KomaMRIBase v0.9.0 `/var/lib/buildkite-agent/builds/gpuci-7/julialang/komamri-dot-jl/KomaMRIBase`
--
[4baa4f4d] KomaMRICore v0.9.0 `/var/lib/buildkite-agent/builds/gpuci-7/julialang/komamri-dot-jl/KomaMRICore`
```

With 1.11 it is:

```
[d0bc0b20] KomaMRIBase v0.9.0
--
[4baa4f4d] KomaMRICore v0.9.0
```

This may be a difference in how `Pkg.develop()` works in Julia 1.11 vs 1.10. I think the pipeline step to develop the packages needs to be done in the KomaMRICore test environment, by adding `--project=KomaMRICore/test` in the code below:

```
julia -e 'println("--- :julia: Instantiating project")
     using Pkg
     Pkg.develop([
         PackageSpec(path=pwd(), subdir="KomaMRIBase"),
         PackageSpec(path=pwd(), subdir="KomaMRICore"),
     ])'
```