@cncastillo 2025-02-05
Hi Ryan! @Stockless  run this PR against master. It seems that it improves the results significantly! 

master (Left) ------------------ This PR (right) 
<p align="center">
  <img src="https://github.com/user-attachments/assets/31f9a331-a454-4cfb-a430-db2952d28a8b" width="45%">
  <img src="https://github.com/user-attachments/assets/d9620b1c-1a08-484c-a403-a9a2b2f95810" width="45%">
</p>

We are currently re-running these with more spins so the differences are more apparent, but thanks for these nice improvements! :smile: 

Do you think you would be able to write a paragraph explaining in simple terms the key aspects that enabled these improvements (compared to array-based)? 

I think it would be really interesting to write a scientific paper with you as co-author. The public of the paper would be MRI experts, which are not really familiar with GPU programming (basically for them writing something in PyTorch is the same as GPU programming). So that paragraph could be used for both the docs, and also the paper.