@rkierulf 2025-02-07
Great! I will add some more info to the docs when I have more time. In the meantime, it would be a good idea for someone with a Nvidia GPU to determine the optimal group size for the precession and excitation kernels by changing the simulation parameters "gpu_groupsize_precession" and "gpu_groupsize_precession". For now, I hardcoded the default group sizes based on what I found to be optimal (32 and 256 for precession / excitation), but if it varies based on the GPU backend then it would be a good idea to change to using a function which dispatches on the backend. As a reminder, the group size should be a multiple of 32 between 32 and 1024, and likely also a power of 2, so I would try 32, 64, 128, 256, 512, and 1024. [Profilng the kernel calls with NSight Compute](https://github.com/JuliaGPU/CUDA.jl/blob/master/docs/src/development/profiling.md) would also be helpful to shed more light on the nature of the kernels (register usage, memory throughput, etc.) and see if anything could be done to make the kernels faster, as well as provide more details that could go in a paper.